+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to Audit.NET and its extensions will be documented in this f
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [PRERELEASE 11.0.4] - 2018-?
+### Added
+- Audit.WebApi: Added GetCurrentAuditScope(this HttpContext httpContext) extension to get the web api audit scope directly from an HttpContext
+
+### Changed
+- Audit.NET.Postgres: Fix the insert command for the Postgres provider (#98)
+
 ## [11.0.3] - 2018-02-12
 ### Added
 - Added request body for AspNet Core Web API 2 via IncludeRequestBody property.